% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi-dx10lookup-data.R
\docType{data}
\name{cfi_dx10lookup}
\alias{cfi_dx10lookup}
\title{CFI ICD-10 Lookup.}
\format{
An object of class \code{data.frame} with 44807 rows and 2 columns.
}
\source{
This was created by Kim et al.
}
\usage{
data(cfi_dx10lookup)
}
\description{
A lookup dataset for CFI ICD-10.
}
\keyword{datasets}
