% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi-dx9lookup-data.R
\docType{data}
\name{cfi_dx9lookup}
\alias{cfi_dx9lookup}
\title{CFI ICD-9 Lookup.}
\format{
An object of class \code{data.frame} with 107 rows and 3 columns.
}
\source{
This was created by Kim et al.
}
\usage{
data(cfi_dx9lookup)
}
\description{
A lookup dataset for CFI ICD-9.
}
\keyword{datasets}
