% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi-pxlookup-data.R
\docType{data}
\name{cfi_pxlookup}
\alias{cfi_pxlookup}
\title{CFI Procedure Codes Lookup.}
\format{
An object of class \code{data.frame} with 90 rows and 3 columns.
}
\source{
This was created by Kim et al.
}
\usage{
data(cfi_pxlookup)
}
\description{
A lookup dataset for CFI Procedure Codes.
}
\keyword{datasets}
