% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi-weightlookup-data.R
\docType{data}
\name{cfi_weightlookup}
\alias{cfi_weightlookup}
\title{CFI Procedure Codes Lookup.}
\format{
An object of class \code{data.frame} with 93 rows and 2 columns.
}
\source{
This was created by Kim et al.
}
\usage{
data(cfi_weightlookup)
}
\description{
A lookup dataset for CFI weights.
}
\keyword{datasets}
