% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-data.R
\docType{data}
\name{id}
\alias{id}
\title{Example ID data.}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{patient_id}{patient_id}
\item{date_of_interest10}{the date of interest, if you were to use only
ICD-10 data}
\item{date_of_interest9}{The date of interest, if you were to use only
ICD-9 data}
}
}
\source{
This was created by the package author.
}
\usage{
data(id)
}
\description{
A dataset with fake patient data, to match the diagnoses, that includes
a date of interest to demonstrate how we can attach these dates and then
subset the data to a specific time window around the date of interest.
}
\keyword{datasets}
