% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepared-data-data.R
\docType{data}
\name{prepared_data}
\alias{prepared_data}
\title{Prepared dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 5 columns.
}
\source{
Built using the packages in this code.
}
\usage{
data(prepared_data)
}
\description{
A dataset which has been prepared using the prepare_data function in this package.
}
\keyword{datasets}
